combined vibes and notes
