+++ app/package.json
@@ -6,7 +6,7 @@
     "postinstall": "npx playwright install",
     "dev": "next dev",
     "playwright-install-deps": "yum install -y alsa-lib atk cups-libs gtk3 libXcomposite libXdamage libXrandr libXScrnSaver libXtst pango libdrm libgbm nss",
-    "build": "cpx \"../vibes/html/**/*\" public/vibes-standalone && next build && tsx src/scripts/generate-ogp-with-static-server.ts",
+    "build": "cpx \"../contents/**/*.html\" public/standalone && next build && tsx src/scripts/generate-ogp-with-static-server.ts",
     "vercel-build": "npm run playwright-install-deps && npm run build",
     "start": "next start",
     "lint": "next lint"
