--- notes/.gitignore
@@ -1,147 +1,0 @@
-*.bundle.*
-lib/
-node_modules/
-*.egg-info/
-.ipynb_checkpoints
-*.tsbuildinfo
-
-# Created by https://www.gitignore.io/api/python
-# Edit at https://www.gitignore.io/?templates=python
-
-### Python ###
-# Byte-compiled / optimized / DLL files
-__pycache__/
-*.py[cod]
-*$py.class
-
-# C extensions
-*.so
-
-# Distribution / packaging
-.Python
-build/
-develop-eggs/
-dist/
-downloads/
-eggs/
-.eggs/
-lib/
-lib64/
-parts/
-sdist/
-var/
-wheels/
-pip-wheel-metadata/
-share/python-wheels/
-.installed.cfg
-*.egg
-MANIFEST
-
-# PyInstaller
-#  Usually these files are written by a python script from a template
-#  before PyInstaller builds the exe, so as to inject date/other infos into it.
-*.manifest
-*.spec
-
-# Installer logs
-pip-log.txt
-pip-delete-this-directory.txt
-
-# Unit test / coverage reports
-htmlcov/
-.tox/
-.nox/
-.coverage
-.coverage.*
-.cache
-nosetests.xml
-coverage.xml
-*.cover
-.hypothesis/
-.pytest_cache/
-
-# Translations
-*.mo
-*.pot
-
-# Scrapy stuff:
-.scrapy
-
-# Sphinx documentation
-docs/_build/
-
-# PyBuilder
-target/
-
-# pyenv
-.python-version
-
-# celery beat schedule file
-celerybeat-schedule
-
-# SageMath parsed files
-*.sage.py
-
-# Spyder project settings
-.spyderproject
-.spyproject
-
-# Rope project settings
-.ropeproject
-
-# Mr Developer
-.mr.developer.cfg
-.project
-.pydevproject
-
-# mkdocs documentation
-/site
-
-# mypy
-.mypy_cache/
-.dmypy.json
-dmypy.json
-
-# Pyre type checker
-.pyre/
-
-# OS X stuff
-*.DS_Store
-
-# End of https://www.gitignore.io/api/python
-
-_temp_extension
-junit.xml
-[uU]ntitled*
-notebook/static/*
-!notebook/static/favicons
-notebook/labextension
-notebook/schemas
-docs/source/changelog.md
-docs/source/contributing.md
-
-# playwright
-ui-tests/test-results
-ui-tests/playwright-report
-
-# VSCode
-.vscode
-
-# RTC
-.jupyter_ystore.db
-
-# yarn >=2.x local files
-.yarn/*
-.pnp.*
-ui-tests/.yarn/*
-ui-tests/.pnp.*
-
-# keep potential upstream patches
-!.yarn/patches
-
-# generated html
-notebook/templates/*.html
-
-# pixi environments
-.pixi
-*.egg-info
